# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/engine": {
      "get": {
        "summary": "List Engines",
        "operationId": "list_engines",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EngineOut"
                  },
                  "type": "array",
                  "title": "Response List Engines"
                }
              }
            }
          }
        }
      }
    },
    "/runners": {
      "get": {
        "summary": "List Runners",
        "operationId": "list_runners",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunnerOut"
                  },
                  "type": "array",
                  "title": "Response List Runners"
                }
              }
            }
          }
        }
      }
    },
    "/game": {
      "post": {
        "summary": "Start Game",
        "operationId": "start_game",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartGameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Games",
        "operationId": "list_games",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameOut"
                  },
                  "title": "Response List Games"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/game/{game_id}": {
      "get": {
        "summary": "Get Game",
        "operationId": "get_game",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stream/game/{game_id}": {
      "get": {
        "summary": "Sse Stream",
        "operationId": "sse_stream",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                },
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/FenEvent"
                    },
                    {
                      "$ref": "#/components/schemas/GameStartEvent"
                    },
                    {
                      "$ref": "#/components/schemas/MoveEvent"
                    },
                    {
                      "$ref": "#/components/schemas/GameEndEvent"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "fen": "#/components/schemas/FenEvent",
                      "game_start": "#/components/schemas/GameStartEvent",
                      "move": "#/components/schemas/MoveEvent",
                      "game_end": "#/components/schemas/GameEndEvent"
                    }
                  },
                  "title": "Response 200 Sse Stream"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stream/live": {
      "get": {
        "summary": "Sse Live Stream",
        "operationId": "sse_live_stream",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/LiveStreamEvent"
                },
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EngineOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "command": {
            "type": "string",
            "title": "Command"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "command",
          "description"
        ],
        "title": "EngineOut"
      },
      "FenEvent": {
        "properties": {
          "type": {
            "type": "string",
            "const": "fen",
            "title": "Type",
            "default": "fen"
          },
          "fen": {
            "type": "string",
            "title": "Fen"
          },
          "ply": {
            "type": "integer",
            "title": "Ply"
          },
          "white_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "White Name"
          },
          "black_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Black Name"
          },
          "moves": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Moves"
          },
          "white_clock": {
            "type": "number",
            "title": "White Clock"
          },
          "black_clock": {
            "type": "number",
            "title": "Black Clock"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "$ref": "#/components/schemas/GameStatus"
          },
          "game_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Id"
          }
        },
        "type": "object",
        "required": [
          "fen",
          "ply",
          "white_name",
          "black_name",
          "moves",
          "white_clock",
          "black_clock",
          "result",
          "status",
          "game_id"
        ],
        "title": "FenEvent"
      },
      "GameEndEvent": {
        "properties": {
          "type": {
            "type": "string",
            "const": "game_end",
            "title": "Type",
            "default": "game_end"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "pgn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pgn"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "GameEndEvent"
      },
      "GameOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "white_id": {
            "type": "string",
            "format": "uuid",
            "title": "White Id"
          },
          "black_id": {
            "type": "string",
            "format": "uuid",
            "title": "Black Id"
          },
          "white_name": {
            "type": "string",
            "title": "White Name"
          },
          "black_name": {
            "type": "string",
            "title": "Black Name"
          },
          "status": {
            "$ref": "#/components/schemas/GameStatus"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "moves": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Moves"
          },
          "fen": {
            "type": "string",
            "title": "Fen"
          },
          "pgn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pgn"
          },
          "white_clock": {
            "type": "number",
            "title": "White Clock"
          },
          "black_clock": {
            "type": "number",
            "title": "Black Clock"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "white_id",
          "black_id",
          "white_name",
          "black_name",
          "status",
          "result",
          "moves",
          "fen",
          "pgn",
          "white_clock",
          "black_clock",
          "created_at",
          "ended_at"
        ],
        "title": "GameOut"
      },
      "GameStartEvent": {
        "properties": {
          "type": {
            "type": "string",
            "const": "game_start",
            "title": "Type",
            "default": "game_start"
          },
          "white_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "White Name"
          },
          "black_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Black Name"
          },
          "game_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Id"
          }
        },
        "type": "object",
        "required": [
          "white_name",
          "black_name",
          "game_id"
        ],
        "title": "GameStartEvent"
      },
      "GameStatus": {
        "type": "string",
        "enum": [
          "playing",
          "ended",
          "aborted"
        ],
        "title": "GameStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LiveStreamEvent": {
        "properties": {
          "game_id": {
            "type": "string",
            "format": "uuid",
            "title": "Game Id"
          },
          "event": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/FenEvent"
              },
              {
                "$ref": "#/components/schemas/GameStartEvent"
              },
              {
                "$ref": "#/components/schemas/MoveEvent"
              },
              {
                "$ref": "#/components/schemas/GameEndEvent"
              }
            ],
            "title": "Event",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "fen": "#/components/schemas/FenEvent",
                "game_end": "#/components/schemas/GameEndEvent",
                "game_start": "#/components/schemas/GameStartEvent",
                "move": "#/components/schemas/MoveEvent"
              }
            }
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "event"
        ],
        "title": "LiveStreamEvent"
      },
      "MoveEvent": {
        "properties": {
          "type": {
            "type": "string",
            "const": "move",
            "title": "Type",
            "default": "move"
          },
          "uci": {
            "type": "string",
            "title": "Uci"
          },
          "san": {
            "type": "string",
            "title": "San"
          },
          "from_square": {
            "type": "string",
            "title": "From Square"
          },
          "to_square": {
            "type": "string",
            "title": "To Square"
          },
          "fen": {
            "type": "string",
            "title": "Fen"
          },
          "ply": {
            "type": "integer",
            "title": "Ply"
          },
          "white_clock": {
            "type": "number",
            "title": "White Clock"
          },
          "black_clock": {
            "type": "number",
            "title": "Black Clock"
          }
        },
        "type": "object",
        "required": [
          "uci",
          "san",
          "from_square",
          "to_square",
          "fen",
          "ply",
          "white_clock",
          "black_clock"
        ],
        "title": "MoveEvent"
      },
      "RunnerOut": {
        "properties": {
          "runner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Runner Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "runner_id",
          "name"
        ],
        "title": "RunnerOut"
      },
      "StartGameRequest": {
        "properties": {
          "white_engine_id": {
            "type": "string",
            "format": "uuid",
            "title": "White Engine Id"
          },
          "black_engine_id": {
            "type": "string",
            "format": "uuid",
            "title": "Black Engine Id"
          },
          "runner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Runner Id"
          }
        },
        "type": "object",
        "required": [
          "white_engine_id",
          "black_engine_id",
          "runner_id"
        ],
        "title": "StartGameRequest"
      },
      "StartGameResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "white": {
            "type": "string",
            "format": "uuid",
            "title": "White"
          },
          "black": {
            "type": "string",
            "format": "uuid",
            "title": "Black"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "white",
          "black"
        ],
        "title": "StartGameResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}